Updated build process for ptII osgi-2-0 jars.
Bug #145 Create update site for Triquetrum

As part of this work, I updated the Jenkins build on
terra.eecs.berkeley.edu so that we delete the ptII osgi-2-0 update
site before copying it over.  In this way, we avoid storing lots
of jar files on terra.  This probably means that each time
we run the build over there, we need to regenerate the tpd file.