Bug #215 Updated aoc vendor to Eclipse Triquetrum.
Initially, the org.eclipse.triquetrum.workflow.aoc.repository
bundle had iSencia as the vendor.  I changed it to
iSencia Belgium NV or something similar.  Erwin correctly
pointed out that as this bundle is in the triquetrum repo,
it should have Eclipse Triquetrum as the vendor.

Some of the bundles in the osgi-2-0 correctly have iScencia 
Belgium NV as the vendor.

Signed-off-by: Christopher Brooks <cxh@eecs.berkeley.edu>